+++ src/num/multind.h
@@ -177,7 +177,7 @@ extern int md_min_idx(unsigned long flags);
 
 #define MD_ACCESS(N, strs, pos, x)	(*({ auto _x = (x); &((_x)[md_calc_offset((N), (strs), (pos)) / (long)sizeof((_x)[0])]); }))
 
-#define MD_STRIDES(N, dims, elsize)	(md_calc_strides(N, alloca(N * sizeof(long)), dims, elsize))
+#define MD_STRIDES(N, dims, elsize)	(md_calc_strides(N, alloca((unsigned long)N * sizeof(long)), dims, elsize))
 
 #define MD_SINGLETON_DIMS(N)				\
 ({							\
@@ -190,7 +190,7 @@ extern int md_min_idx(unsigned long flags);
 #define MD_SINGLETON_STRS(N)				\
 ({							\
 	int _N = (N);					\
-	long* _dims = alloca(_N * sizeof(long)); 	\
+	long* _dims = alloca((unsigned long)_N * sizeof(long)); 	\
 	md_singleton_strides(_N, _dims); 		\
 	_dims; 						\
 })
